wip: Refactored editor
