+++ src/src/pages/release-notes/editor/tiptap/components/tiptap-templates/simple/simple-editor.scss
@@ -37,7 +37,7 @@ body {
 }
 
 .simple-editor-content {
-  max-width: 640px;
+  // max-width: 640px;
   width: 100%;
   margin: 0 auto;
 }
